models.Member.AttendancePerTechnique now returns all techniques
This method now adds all the techniques that the member has not attended
so that the model function now returns a valid entry for all techniques.
This simplifies the views.

One consequence appears to be it is now not possible to filter or get
based on the values of the query set. It appears to have something to do
with the union. The ordering seems to work, though.