Add Technique.MemberAttendance method
I was having trouble combining the two query sets for the techniques
with members and those with a count of 0, so instead I created an
iterable with the two query sets. This results in an object that
itself is not a queryset but can be iterated over to get the Member
objects.

For completeness, I changed the Member.AttendancePerTechnique method
to also return an iterable. This might solve problems later if
changes in the query make things uncombinable.